chore: move hook to scrollHelpers
